Partial progress
Current best effort - using :browser mode on a separate :dev build,
compiling outside frontend/src and included as an "external" in webpack.

That's not working, though, because the resulting browser bundle:
- can't be loaded as a required webpack import, because it wants `this`
  to be `window`, not `module.exports == {}`.
- can't be loaded as a "script" external, because webpack loading
  continues while the script is loaded in the background, and the "var"
  imports that want to look up eg. window.metabase.foo.bar fail.

I'm not sure what the path forward is. I'm going to experiment later
with using `:npm-module` but with the code outside `src/` like that,
using the webpack alias to find it in `target/cljs` or whatever.

I think that will work and avoid some of the janky HMR overlap, but I
expect shadow-cljs hot reloads and the REPL to fail the same way as
before - they're not reloading into the right context, I think because
they don't understand how webpack loaded the original code.

That can probably be fixed on the shadow-cljs side by fixing
`:npm-module` or adding new output options to it to make it play nicer
with webpack etc.